Don't create debian/changelog when it exists

+++ helpers/build-debs.sh
@@ -34,32 +34,36 @@ for dir in "${dirs[@]}"; do
   echo "==================================================================="
   echo "-> ${dir}"
   pushd "${dir}"
-  # Parse the Source name
-  sourcename=`grep '^Source: ' debian/control | sed 's,^Source: ,,'`
-  if [ -z "${sourcename}" ]; then
-    echo "Unable to parse name of the source from ${dir}"
-    exit 1
-  fi
-  # Let's try really hard to find the release name of the distribution
-  distro_release="$(source /etc/os-release; printf ${VERSION_CODENAME})"
-  if [ -z "${distro_release}" -a -n "$(grep 'VERSION_ID="14.04"' /etc/os-release)" ]; then
-    distro_release='trusty'
-  fi
-  if [ -z "${distro_release}" ]; then
-    distro_release="$(perl -n -e '/VERSION=".* \((.*)\)"/ && print $1' /etc/os-release)"
-  fi
-  if [ -z "${distro_release}" ]; then
-    echo 'Unable to determine distribution codename!'
-    exit 1
-  fi
-  set_debian_versions
-  cat > debian/changelog << EOF
+  # If there's a changelog, this is probably a vendor dependency or versioned
+  # outside of pdns-builder
+  if [ ! -f debian/changelog ]; then
+    # Parse the Source name
+    sourcename=`grep '^Source: ' debian/control | sed 's,^Source: ,,'`
+    if [ -z "${sourcename}" ]; then
+      echo "Unable to parse name of the source from ${dir}"
+      exit 1
+    fi
+    # Let's try really hard to find the release name of the distribution
+    distro_release="$(source /etc/os-release; printf ${VERSION_CODENAME})"
+    if [ -z "${distro_release}" -a -n "$(grep 'VERSION_ID="14.04"' /etc/os-release)" ]; then
+      distro_release='trusty'
+    fi
+    if [ -z "${distro_release}" ]; then
+      distro_release="$(perl -n -e '/VERSION=".* \((.*)\)"/ && print $1' /etc/os-release)"
+    fi
+    if [ -z "${distro_release}" ]; then
+      echo 'Unable to determine distribution codename!'
+      exit 1
+    fi
+    set_debian_versions
+    cat > debian/changelog << EOF
 $sourcename (${BUILDER_DEB_VERSION}-${BUILDER_DEB_RELEASE}.${distro_release}) unstable; urgency=medium
 
   * Automatic build
 
  -- PowerDNS.COM AutoBuilder <noreply@powerdns.com>  $(date -R)
 EOF
+  fi
 
   fakeroot debian/rules binary || exit 1
   popd
